Cleanup + use CocoaPods

+++ VybeTen/VYBLabel.m
@@ -22,7 +22,7 @@
 /* Drop shadow on the text */
 - (void)drawTextInRect:(CGRect)rect {
     CGSize myShadowOffset = CGSizeMake(0, 0);
-    float colorValues[] = {0, 0, 0, .8};
+    CGFloat colorValues[] = {0, 0, 0, .8};
    
     
     CGContextRef myContext = UIGraphicsGetCurrentContext();
